add std.atan2 as a builtin
It can be implemented as jsonnet code in std.jsonnet, but making it
as a builtin has the advantage of relying on the probably-well-tested
code of the C / C++ standard library. Numeric functions can have
subtle or non-obvious boundary cases and calculation precision
issues in naive implementations, so even though atan2 _seems_ simple
enough, I'd prefer to use a well tested implementation.